All tests passed

+++ c/all-your-base/all_your_base.h
@@ -5,6 +5,6 @@
 
 #define DIGITS_ARRAY_SIZE 64
 
-int64_t rebase(int8_t digits[DIGITS_ARRAY_SIZE], int16_t input_base, int16_t output_base, int64_t input_length);
+int8_t rebase(int8_t digits[DIGITS_ARRAY_SIZE], int16_t input_base, int16_t output_base, int8_t input_length);
 
 #endif
